requirements.txt
Flask
Flask-Login
Flask-WTF
psycopg2-binary
gunicorn
boto3